Struct socks::Socks4Listener

pub struct Socks4Listener(_);

A SOCKS4 BIND client.

Methods

impl Socks4Listener

fn bind<T, U>(proxy: T, target: U, userid: &str) -> Result<Socks4Listener> where T: ToSocketAddrs, U: ToTargetAddr

Initiates a BIND request to the specified proxy.

The proxy will filter incoming connections based on the value of target.

fn proxy_addr(&self) -> Result<SocketAddr>

The address of the proxy-side TCP listener.

This should be forwarded to the remote process, which should open a connection to it.

fn accept(self) -> Result<Socks4Stream>

Waits for the remote process to connect to the proxy server.

The value of proxy_addr should be forwarded to the remote process before this method is called.