Struct socks::Socks4Stream

pub struct Socks4Stream { /* fields omitted */ }

A SOCKS4 client.

Implementations

impl Socks4Stream

pub fn connect<T, U>(proxy: T, target: U, userid: &str) -> Result<Socks4Stream> where
T: ToSocketAddrs,
U: ToTargetAddr,

Connects to a target server through a SOCKS4 proxy.

Note

If target is a TargetAddr::Domain, the domain name will be forwarded to the proxy server using the SOCKS4A protocol extension. If the proxy server does not support SOCKS4A, consider performing the DNS lookup locally and passing a TargetAddr::Ip.

pub fn proxy_addr(&self) -> SocketAddrV4

Returns the proxy-side address of the connection between the proxy and target server.

pub fn get_ref(&self) -> &TcpStream

Returns a shared reference to the inner TcpStream.

pub fn get_mut(&mut self) -> &mut TcpStream

Returns a mutable reference to the inner TcpStream.

pub fn into_inner(self) -> TcpStream

Consumes the Socks4Stream, returning the inner TcpStream.